Fix CD-ROM ISO creation for VMware jobs on Windows #1553
For a VMware VM universe job, if the user used the vm_cdrom_files option
to have Condor create a CD-ROM image for the VM, the vm gahp would fail
if the vm already had a cd-rom image file associated with it.
Specifically, if the .vmx file had a line like this:

  ide1:0.deviceType = "cdrom-image"

The vm gahp expected the value to be "atapi-cdrom" or "cdrom-raw".
Now, "cdrom-image" is accepted as well. The result is that the existing
cd-rom image is replaced with the one Condor is asked to create.