Fix vm_no_output_vm job attribute in vm gahp #1564
The vm gahp was attempting to remove job files as the wrong user,
allowing the files to be transferred back to the submit machine when
they shouldn't be. The most noticeable case of incorrect transfer is
when vm_no_output_vm is set to true in the submit file.

Now, the vm gahp does the removal as the executing user, rather than as
the condor user.