Reverse direction of alive messages between schedd and startd #1420 #671
Claim keep-alives now default to being sent by the startd, rather than
the schedd. The default value of STARTD_SENDS_ALIVES is now 'peer',
which means the daemon's behavior is based on the version of its peer.
If the peer is older the 7.5.4, the old default of schedd sending alives
is used. Otherwise, the new default behavior is used. The parameter can
still be set to True or False, in which case all daemons (new and old)
will always use the specified behavior.