Remove Scheduler/Submitter ad type hack in CondorQuery #1581
When querying for submitter ads, use ad type "Submitter" instead of
"Scheduler". This is one step in eliminating an old compatibility hack
where submitter ads had a MyType of "Scheduler".

As a result, condor_status and condor_q will not see submitter ads
advertised by a pre-7.3.2 schedd to a pre-7.3.2 collector. The
negotiator will continue to see these ads and negotiate appropriately.