Optimize query processing in collector #1579
Now that queries sent to the collector don't use TARGET. scoping, we can
be more efficient in applying the filter expression to ads. We no longer
use a match ad to do a one-way match. We cache a pointer to the
expression and evaluate it directly against each ad. Also, we only check
the MyType if we're searching more than one type of ad.