Build user-log testing binaries during build, rather than test.
About 15 tests use a few user-log read/writing utilities. Before, these
utilities were built as part of the NMI test run. This caused these
tests to fail on many of the cross-test platforms, due to C++ ABI
differences when trying to link libcondorapi.

Now, the utilities are created during the build and placed in
testbin_dir.