Schedd now writes the submit event in the user log #1441
Previously, condor_submit wrote the submit event, but the schedd (or the
gridmanager) wrote all other events. The schedd even wrote the submit
event to the event log. This can cause synchronization problems, such as
having execute/terminate events written first.

Now, the schedd writes the submit event along with all other events. If
the schedd or condor_submit detect that they're talking to an older
peer, they revert to the old behavior (condor_submit writes the submit
event).

Also, the submit event in the event log now contains the same
information as the submit event in the user log. Previously, it lacked
the schedd sinful string and any submit event notes.