Fix invocation of priv-sep switchboard in vm gahp. #1611
When the VM gahp invoked one of its helper scripts via priv-sep, it
launched the switchboard with execvp(), giving it the basename of the
switchboard binary (i.e. not the full path to the file). As a result,
the switchboard binary was looked for in the path. This could result in
the binary not being found or a wrong one being used.

Now, the full path of the switchboard binary is used.