Fix vm gahp script callouts when using priv-sep. #1612
When the vm universe is used with privilege separation, the vm files
spend some of their lives owned by condor and some time owned by the
job user. When the files are owned by condor, any callouts to the vm
gahp's helper script need to be done as condor, not the job user.