Submit shouldn't pad DiskUsage for expanding vm image. #1624
When calculating the DiskUsage attribute for a vm job, condor_submit
would add an extra 100MB because the vm image may be a sparse disk and
thus expand during execution. It does no padding for other types of jobs,
which usually write output files. The amount of padding seems arbitrary.
So I'm removing the padding for vm universe jobs.