This fixes a problem in Condor's file transfer semantics.
The problem is that if someone specifies transfer_output_files and is
using file transfer, then if the job doesn't produce all of the files
in question and exits (suppose by segfault), Condor will put the job
on hold. In this circumstance there will be no record of how the job
terminated. This is problematic because: it is hard to debug simply by
looking at the job ad of the held job, and the user policy expression
like the on_exit_* and periodic_* expressions can't be used to make a
judgment about the job.

I fixed it so that when the file transfer fails because not all of
the files were present, we "mock terminate" the job via a new remote
system call from the starter to the shadow which happens after the
failed transfer, but before the starter tells the shadow to put the job
on hold due to the failed transfer. This catches things like the job
segfaulting--which presumably is why it didn't produce all of its output
file. This new remote system call in the starter will only happen for
7.4.4 shadows or later. A mock terminate simply means to commit into the
schedd's authoritative copy of the job ad with how the job terminated,
but don't write a terminate event into the job log or otherwise act as
if the job has been terminated. Even though we mock terminate the job,
the job will still go on hold in these situations. The job ad will still
contain the hold information and additionally the termination information
of the job.

This is for ticket #1587 and implements only Number 1 in the list.