Add recognition of new STAGING_OUT job status. #1641
This value for the JobStatus attribute is to be used when the job's
output files are being staged back to the submit machine. Currently,
jobs are marked as RUNNING during this time.

The STAGING_OUT status is now recognized everywhere and treated the same
as the RUNNING status. Nothing changes a job's status to STAGING_OUT
yet.