Start using new JobStatus STAGING_OUT. #1641
The starter now sends a job ad update indicating the job has exited to
the shadow before starting the transfer of output files. When the shadow
receives this update, it changes the job's status to STAGING_OUT in the
schedd. This change only affects vanilla, java, and vm universe jobs.