Make gridmanager detection of gram version more cautious. #1653
Concluding that a gt2 server is gt5 can have disastrous results. Now,
if the gridmanager is unsure of a server's version, it proceeds as if
it's gt2. If the version later turns out to be gt5, it transistions
appropriately, signaling any jobs held up by gt2 throttles.

Also, there's a new config knob named GRAM_VERSION_DETECTION that can be
used to disable gram version detection. If set to false, then the grid
type given in the GridResource attribute of the first job seen for a
given resource is trusted for the gram version.