Fixed getrlimit() in the stduniv.
The 64-bit version was broken by old crusty compatibility code in cedar
from when the fields of the rlimit structure changed from signed to unsigned.

The 32-bit version was broken because we needed to use SYS_ugetrlimit syscall
number instead of the older, and 2gb limited, SYS_getrlimit.

This is for ticket #1654.