Schedd controls who sends keep-alives. #1652
STARTD_SENDS_ALIVES is an error-prone knob, especially when flocking. If
the submit and execute machines don't have it set the same, then you can
end up in a situation where neither daemon is sending keep-alives, and
all jobs die an early death.

Now, the schedd decides who sends the keep-alives and informs the startd
of its choice. Both startd and schedd still act in the old way when
interacting with older peers.