Revamp how submit splits attributes between cluster and proc ads. #1650
Now, when condor_submit builds up each job ad in a cluster, all
attributes are inserted into the local copy of the ad. When submit sends
the ad to the schedd, it omits attribute/value pairs that are identical
to what's in the cluster ad.

Previously, the local copy of the ad would only contain attributes that
differed from the cluster ad. This caused problems for the
VMware-related code, which reads the previously-set value of
TransferInput from the job ad and adds additional files to it.