The Problem: When there is an error, we create a new row in the error_sqllog table and place the ENTIRE log file into the newly created row. Under conditions where the condor_quilld is down for a long time, the log file may have grown to hundreds of megabytes.  Additionally, if a stable error rate is being produced by quill, like say thousands of jobs in schedd_sql.log have a type constraint error for a jobad attribute and those jobs aren't leaving the queue anytime soon, you'll get thousands of copies of the log file inserted into the database when condor_quilld is restarted. This will easily fill most reasonable disk volumes. This scenario actually happened at a customer's site and it took days and finding a *huge* disk volume to use to recover from it.
The Solution:
I ifdefed out the portion of the code which reads the schedd_sql.log
file and inserts it into the LogBody column in the newly created row
associated with the sql error.

Why:
A better fix isn't being produced since, A) quill is going dark,
B) we know of noone using this functionality, and C) this is a
stable series and a small patch needed to be produced.

Also I stumbled upon and fixed a close() of an uninitialized variable
during error paths.