Fix TmpDir long path name test on AIX. #1239 #1684
AIX's getcwd() can't handle paths longer than 1023 characters. It will
truncate the path and return success. The TmpDir unit tests now use a
shorter set of paths for AIX to avoid this problem.

This change also restores the longer pathnames in the unit tests when on
platforms with a PATH_MAX larger than 1024.