Fix daemon-core http get for non-linux unix platforms. #1701
Function serve_file() in src/condor_daemon_core.V6/soap_core.cpp  calls
libc function realpath() incorrectly on non-linux unix machines. It
relies on a glibc extension that will malloc space for the function's
answer if NULL is passed as the second argument. On AIX and Solaris,
this causes the call to fail. On Mac OS X 10.4 and 10.5, this causes the
program to crash. Mac OS X 10.6 adds the extended behavior.

Now, we provide a buffer for realpath() to write the result into.