Retry purge of cancelled cream jobs in gridmanager. #1700
When the gridmanager needs to cancel a Cream job, it issues a cancel
command, followed by a purge command. If these commands are too closely
spaced, the purge can fail because the Cream server hasn't finished
processing the cancel. The server will return the error "job status does
not match".

In this situation, we now retry the purge command after a short delay.