Fixed an obscure bug with getcwd() in condor_getcwd() that our unit tests found. Under Solaris, if getcwd() is called in a directory which is larger than PATH_MAX then getcwd() returns NULL and ERANGE regardless if the buffer and buflen passed into it are plenty big enough to hold the path or not--forever. Since condor_getcwd() tries to find a large enough path, it'll eventually consume all of the memory of the process and kill the process or cause other havoc on the machine.
The kinds of havoc which can be caused are exhausting the virtual memory
space on Solaris (which is also space in /tmp), or causing the OOM to
wake up and go on an indescriminate killing spree in Linux, etc, etc, etc.

I worked around this bug so that if we try to allocate memory which
is larger than 20Megs in order to put a path into it, we declare that
getcwd() has failed and bail in trying to find a larger path. I suspect
other OSes have similar problems since this is a bit of a problem area
for getcwd() in general.

This is for ticket #1716.