Addressed problem of too many entries in the spool directory. #1706 This problem affects spooling of more than 16k jobs to an ext3 directory, because we run into the 32k link limit (two spool directories per job).
Avoided this by adding a hierarchy for the per-job directories and files
stored in the spool.

Also added spool version info to improve the upgrade/downgrade process.
Now we can tell what format the spool is in when we start up and we
can record the minimum version that should try to read the spool
in case the user downgrades.