Fixed a bug where we didn't process the return value of open() properly with regards to if the command had shell syntax in it or not.
Fixed a bug handling $? which should be checked after the close.

Got rid of a sync which hurts the submit machine too much. (like 15+ minute
waits).

Fixed the log_cmd function to use a subshell for whatever it runs for
collecting the stderr and stdout into one stream.

Now if submitting a build fails in "usual" ways, like we didn't have a
condor_nmi_submit program, or some other thing, we skip that build and try the
next one.