This checkin implements two features:
1. If an unknown platform identifier is presented to --platforms, the
'default_minimal_platform' platform configuration is chosen for it out
of the %submit_info hash table in SubmitInfo.pm. No attempt is made to ensure
that this unknown platform is actually available in the batlab.

2. A platform supplied to --platforms can be forced to use any
other legally defined platform configurations in SubmitInfo.pm. The
new syntax for this is platform_name#platform_key. An example:
--platforms=x86_64_rhap5#x86_rhap_5 means to build and test
on an x86_64_rhap_5 machine, but instead use the platform
configuration for x86_rhap_5. You can choose this as well:
x86_64_rhap5#default_minimal_platform.

Specified platform keys MUST exist in the %submit_info database in
SubmitInfo.pm.

A build and test set may use a platform_name#platform_key designator.

Specifying a build and test set name with a platform_key is undefined and
will result in an error:

--platforms=official_ports#minimal_default_platform => ERROR

This is for ticket #1722.