Change spaces to tab in imake rule for Makefile target.
The commands to create a Makefile target must be preceeded by a tab. Our
imake rule had 8 spaces instead. With older versions of the
preprocessor, it appears the spaces were being changed to a tab for us.
On fedora 12 (gcc 4.4.4), it's keeping the spaces, causing the build to
run incorrectly.