Fix Cream external on older Mac OS X platforms.
One of the Cream source files declares 'extern lnog timezone', which is
provided by libc on linux. On Mac OS X, this conflicts with a function
of the same name. Cream doesn't actually use this variable, so we can
just remove the declaration.

This doesn't change how the code builds on other platforms, so no
version bump for the external is needed.