Allow the string "NULL" for m_instanceName.
We were checking for m_instanceName == NULL, but this wasn't
handling the case where we re-started condor (and hence
parsed a gridresource m_instanceName as the string "NULL").
This would lead to all jobs being stuck in HOLD when
condor was restarted.  By handling the case where m_instanceName
is the "NULL" string, we can now successfully restart
condor.

Signed-off-by: Chris Lalancette <clalance@redhat.com>