Fix counts of running scheduler and local universe jobs. #1761
In the schedd, jobs marked for holding or removal which have yet to be
killed are treated inconsistently for LocalUniverseJobsRunning and
SchedUniverseJobsRunning. Between calls to count_jobs(), they're
included in the count. In count_jobs(), they're not included. This can
lead to the counts falling below zero when a job exits. This causes the
schedd to EXCEPT for the local universe case.

Now, these jobs are always included in the running job counts.