Fix sporadic failures of lib_auth_protocol-negot tests.
These tests try all combinations of encryption settings between client
and server (NEVER, OPTIONAL, PREFERRED, REQUIRED). They setup a personal
Condor, run a client tool, and see whether encyrption is used, not used,
or the security negotiation fails. For the NEVER-REQUIRED combinations,
the daemons can't startup because they can't talk to the master. The
test script can end up waiting forever for the daemons to create a
.address file.

Now, only the tools get the client-side encyrption settings for the
test. The daemons use the same settings for client- and server-side
communications, so they can always talk to each other.