Prevent gridmanager crash if gahp misbehaves on startup. #1695
If a gahp server doesn't print a proper version banner or respond to
the COMMANDS command, the gridmanager's GahpServer object is left in a
partially-initialized state. In particular, m_commands_supported is left
as NULL, which can result in a crash if higher-level code tries to issue
any commands to the gahp.

Now, the failed startup is remembered. GahpClient::isStarted() returns
false and subsequent calles to Startup() immediately return false.