Removed special case NEGOTIATOR_MATCH_LOG code from dprintf_config and default in param_info.in, #61
NEGOTIATOR_MATCH_LOG is a log of the form <SUBSYS>_<LEVEL>_LOG, a
feature of dprintf to provide a sink for D_MATCH. The special case
code is unnecessary because the intent of #61 is to avoid EXCEPT from
lack of a _LOG param and NEGOTIATOR_MATCH_LOG is only used if it is
present. The removal from param_info.in is because its presence meant
it was impossible to turn off the special logging.