Fix procd problems #1785 and solaris 11 /dev/null problem #1788
starter would error is user created symbolic link to nowhere in
the sandbox.  Ignore symbolic links when changing owner.  This is
ok because symbolic links are 0777 and the directory owner controls
removal.

If you'd call fopen_wrapper to open /dev/null for writing it would
fail on solaris 11 as it returned ENXIO for an error instead of
EEXISTS when creating a file using O_EXCL and O_CREAT since /dev/null
is a sym link and solaris 11 is stupid.