Fix vm_gahp callouts for xen/kvm. #1803
Due to a bug in a code refactoring for 7.4.4, script callouts from the
vm_gahp that should happen as root instead happen as the executing user.
However, none of the callouts need to happen as root, since all
interactions with libvirt occur via the client library now. Condor 7.5.4
is also affected.

We'll do the following changes now. Additional changes can wait for
later.
* Fix systemCommand() to perform callouts as root when requested.
* Remove check for uid root from condor_vm_xen.sh.
* Remove call to virsh from check routine in condor_vm_xen.sh.