Make MyType and TargetType optional in ClassAd wire protocol. #1164
In new ClassAds, MyType and TargetType are ordinary attributes and not
even required in an ad. However, in the wire protocol inherited from old
ClassAds, values for these attributes are always transmitted. If the
sender didn't set values, then '(unknown type)' or an empty string was
sent. Thus, these attributes are always set on the receiver.

Now, the sender always sends an empty string if the attributes weren't
explicitly set. The receiver doesn't define these attributes if the
transmitted value is an empty string or '(unkonwn type)'. This allows
these attributes to be undefined in the receiver if they don't have a
real value in the sender.