This is the hand merge of the OSG procd fork into mainline Condor.
As of this commit, the condor_procd can function both as the procd for
the Condor system and also for the OSG process family tracking system.
This feature concerns itself when the USE_GID_PROCESS_TRACKING feature
is enabled (or -G <mingid> <maxgid> is supplied to the procd).

In the OSG model, the procd is run in a standalone mode and told what gid
(out of a known valid range) to associate with what process family. Also
the procd will not die if the root pid vanishes. An interesting tidbit
of info in the OSG model is that there maybe X number of procds on a
single machine, monitoring processes, but all of them are using the A to
B range of GIDs. The decision of which gid to use for a new process is
decentralized in that gidd_alloc is used to atomically (well, not really,
there is a race which needs fixing later) find an unassociated gid in the
specified range and not being used by any process on the machine. This
gid is used by the external control scripts controlling the procd with
procd_ctl to associate the gid with the right pid. This decentralization
of gid association is a key thing to realize in the OSG model.

In the Condor model, the procd will self-assign GIDs out of a pool to
the process family. The procd will die if the root pid it is monitoring
in happens to vanish.

When a condor_procd is tracking pids with gids, it can ONLY exist in the
state of self-allocation or externally associated. At this time it is coded
such that those two states are mutually exclusive.

Work done:

+ Added gidd_alloc. Sadly, it is linux only for now. It doesn't have to
be, though. Getting the right prototype for setgroups() without breaking
standard universe (see the ocmments in the code) is the trick which I
didn't have time to figure out.

+ There is now the notion of tracking by allocated gid, or tracking by
associated gids. There are two code paths in the proc_family_client/server
dealing with each concept. However, when it comes to the family
monitor object, there is only one code path. The internal GIDPool upon
construction is told which method it should use. The group-tracker object
just uses the single add_mapping interface to the GIDPool and assume the
GIDPool will do the right thing with it given what mode the GIDPool is
operating in.

+ Added new command line arguments to the condor_procd which allow it to:
accept a root pid to follow, decide to except or not if said root pid
exits, determine if the procd should self-allocate gids when tracking
with gids, or if it knows that an external tool (like the procd_ctl)
will associate a gid with the pid later.

These are:
 -P <pid> :: specify a root pid to follow and do not exit if it dies.
 -E :: when used with -G, it means to NOT self-allocate gids

Part of this change was fixing up the conditions under which the procd
dies if the root pid family (known as the "watcher pid") goes away.

+ Updated a bunch of code internal to the proc_family_monitor object to
use the lookup_family() internal function call.

+ Changed the procd_ctl tool to use TRACK_BY_ASSOCIATED_GID and the
subsequent code path in the proc_family_client object. Added the ability
to specify the procd_address file on the command line which would override
the config files specification of it. Updated the usage information. It
used to (inexplicably) only compile on windows, but not it'll compile
everywhere.

+ Altered/added some debugging messages to make the procd log output a little
more sane.