Support launching DaemonCore children without a UDP command port. #1814
This is used by the schedd when launching shadows in unix.  Previously,
it created a UDP port for the shadow, and then the shadow closed it
when it started up (in order to save memory).  This patch improves
the situation by never creating the UDP socket in the first place.
The advantage is that the schedd knows explicitly not to try to
communicate with the shadow via UDP.