Allow destroy to fail on cleanup.
On stateless clouds (like ec2) the call to destroy will fail here
because "stop" and "destroy" are basically both the same operations, so
this does a double destroy.  On other clouds this destroy call will be
needed to ensure the instance is completely removed and make the
provider act like a stateless provider.

Signed-off-by: Ian Main <imain@redhat.com>