Have begun a simple client and server that follow the protocol on a basic level. Currently working is a partial client that does the following:
     * Opens and determines the size of a file
     * Connects to a server using TCP
     * Sends the SIF frame and a simplified parameter set
     * Reports any errors encountered with any of the above

The next steps will be to finish the basic client by adding
the transfer and teardown phases. Then work can begin on a
simple server program. Together they should allow a file to
be sent and saved on a remote machine.

Additionally, the project uses GNU makefiles - work will need
to be done to use the Condor standard buildsystem. But that can
come later.

 -- Nathan Mitchell