Finished writting the the Server's accept routine
Finished the Client/Server Negotiation phase
   - Its really hacky and lacks more than it contains, but it 'works'

Added some minor tweaks around the code in general
   - Most importantly added standard padding to the frames to
     bring them all into 32 byte sized packages. This allows for
     a single cftp_frame type to be used for receiving and sending
     which then can be casted into the actual type of frame when needed

Fixed up the Makefile to make it cleaner and easier to add to.

Will be working on the Server's transfer section next. Hopefully the next major
commit will have the Server being able to save a file to disk!