Do not let tools hammer DNS
condor_advertise seems to hitting DNS hard.  With this change, we get exactly
one call to gethostbyname() from condor_advertise, and presumably any other
tool that uses DNS caching. This is #1805

This patch looks large, but a lot of it is just whitespace changes from
changing indentation. So one must look carefully at the patch