Order the tasks on the dashboard details page by start time instead of alphabetical ordering.  This is a more logical view.
This fix isn't perfect because when multiple hosts are displayed on a page
(such as the test results for one run id) the fields are sorted by the
start times on one of the hosts.  Any fields not present for that host will
be at the end, after the NMI post_all step.  But this is a minor issue and
this fix looks to be an improvement.