Fix some bugs in attribute updates to the gridmanager. #1790
In BaseJob::JobAdUpdateFromSchedd(), we need to distinguish between full
copies of the job ad from the schedd and ads that only contain modified
attributes. Otherwise, we can end up in an endless loop of retrieving
the full ad and trying to write all of the values back to the schedd.

Added optional parameter to MergeClassAds() to mark merged attributes as
clean. This prevents unnecessary attribute updates from the gridmanager
to the schedd.

Fix leaked ClassAd returned by DCSchedd::clearDirtyAttrs() when called
in doContactSchedd().