Added a lightwieght SHA1 hash library to perform file hashing. The license is free, so there shouldn't be any problems, but it can swapped out of needed (or if something better comes along).
Completed a very basic version of the transfer phase. The client and server
now will transfer a file! There was much rejoicing. Next will be adding
the teardown phase where the server will ensure the the file transferred
successfully by comparing hash values.

Once this is done, the prototype will be done and we can proceed to a
code review.