cmd_q_shows_sub test was failing, as it had a bogus executable
Of all the tests, this was the only one using CondorTest::RunTest2

While RunTest2 was a pleasant idea (passing a perl hash representing
the submit file instead of another file), having only one out of hundred
test use it is not.  I changed this to use the same CondorTest::RunTest
that everyone else uses.