Fix bug that removes running jobs instead of holding them. #1900
The shadow was unconditionally changing the job's state to
TRANSFERRING_OUTPUT when it was told by the starter that the job was no
longer running, regardless of why the job was no longer running. If the
job was held (e.g. by condor_hold), the shadow would still change the
status to TRANSFERRING_OUTPUT once the starter killed the job. This
caused the schedd to remove the job when the shadow exited.