Add new grid-type ec2, which implements the EC2 Query interface. #1821
Add code to condor_gridmanager and condor_submit for a new ec2
grid-type. It's a copy of the amazon grid-type, which it will replace.
The only difference is a change in the names from 'amazon' to 'ec2', and
a change in the names of the job attributes for the authentication
files.