Added param NEGOTIATOR_UPDATE_AFTER_CYCLE, defaults to False.
When true, it will force the Negotiator to publish an update ad to the
Collector at the end of every negotiation cycle. This is very useful
if monitoring LastNegotiationCycle statistics. Without this, when
NEGOTIATOR_INTERVAL * NEGOTIATION_CYCLE_STATS_LENGTH < NEGOTIATOR_UPDATE_INTERVAL
statistics can be missed. Additionally, it is useful to poll
LastNegotiationCycle*0 stats, which will easily miss cycles.