Safeguard Deltacloud credentials #1940
The Deltacloud 'password' is the secret credential required to
authenticate with the cloud service that the Deltacloud service is
fronting. It needs to be protected from other users. Thus, it shouldn't
appear in the job ad. Now, it's kept in a file and the job ad contains
the path to that file.