Add condor_rmdir a new windows utility program that behaves more like the linux rm -rf tool.  This is in support of #1877.
This new program is intended to be used by the Directory object
instead of the windows built-in rmdir.  Unlike rmdir, it can
be told to continue if it fails to delete a file, and to take
ownership and remove DACLs from files and directories if necessary
as part of the deletion process.

The program is stand-alone and contains it's own makefile, although
the condor build process does not use it.  It also has it's own
(tiny) version of the c-runtime so that it won't require a version
of MSVC runtime to be installed in order to work.