Fix schedd writing to user log for remote jobs with unspooled files. #1985
For remotely submitted jobs, once spooling of input files is complete,
the schedd rewrites the job ad so that file paths are relative to the
spool directory. One problem with this is that any user log events
written before the files are spooled are written to the wrong path. This
happens for the submit event and, if spooling doesn't occur, for the
abort event.

To solve this, we now do the modifications of the job ad when the job ad
is initially submitted (in CommitTransaction()). The modifications are
done as a separate job queue log transaction. On startup, the schedd
will redo the rewrites if it detects any jobs that need them.

The code to write the submit event has been changed to mirror the code
used to write other log events.

condor_submit no longer transfers the user log as an input file. As a
result, the schedd always starts a fresh user log in the spool directory
for remote jobs. This user log is transferred as an output file. At
present, it will overwrite any existing user log file on the submitting
machine.