Cosmetic and usability cleanup to condor_q -bet output.
This is a fix for confusing output introduced in an earlier commit
addressing #1670.  Red Hat Bugzilla 692140 is also related.

The earlier fix caused condor_q's analysis support to short-circuit
when jobs have not yet been considered for matchmaking, since this often
corresponds to a particular case of a job being rejected for "unknown
reasons."  However, that behavior can be confusing, since users would
generally like to know why the job would have been matched (if it has
already been through a negotiation cycle).  The behavior introduced in
this commit only short-circuits analysis in the cases in which the job
has not yet been through a negotiation cycle and a machine would be
available to run it (corresponding to the "unknown reasons" cases).
In other cases, potentially-useful information about why occupied
machines rejected the job is still available.

===GT:Fixed=== #1670
===BZ:Fixed=== RHT 692140