Revert "Restore libcondorzsyscall.a.  #1783"
This commit broke the standalone checkpointing of compressed checkpoints
on 32-bit machines. I tested without this patch and standalone checkpoints
started to work again.  gthain and I are reverting this patch so that
at least the head of the V7_6-branch is functional.

This does mean though, is that if 7.6.0 gets released with this reversion
and the real problem hasn't been fixed yet, then new stduniv programs
will use uncompressed checkpoints.

This reverts commit 82552c30a50ab2e45f34a3c8c1118d226128a2ad.