Don't create a spool directory for every job. #2069
Code changes for #1985 accidentally caused a spool directory to be
created for all jobs that need a submit event log to be written by the
schedd (either in the user log or the global event log).
Here, we restore the previous behavior of only creating a spool
directory when required (i.e. when a job is spooling input files to the
schedd).