Fix build flags for Globus external. #2187
With each external being installed in a different directory, the CFLAGS
and LDFLAGS for Globus need to be changed. Otherwise, it uses the system
openssl half the time and the externals version the other half of the
time. This was causing Condor to crash during GSI authentication.

Also, fix the directory containing the voms libraries on Mac OS 10.6.