Always dynamically link libvirt for vm gahp, take 2. #2083
In order for the vm-gahp to work on most or all linux distros, we need
to always use the version of libvirt installed on the executing machine.
This means always dynamically linking. Fortunately, all versions of
libvirt that we care about are library-compatible. So which version we
dynamically link with doesn't matter.

This means when libvirt isn't available on the build machine, we build
and link against the shared libvirt library in our externals.

Side effects: We now use the OS-provided libxml2, except on rhel3, whose
version is too old.