When compiling this ancient crusty version of krb5 (1.4.3-p0) on RHEL 6, there were two problems:
Problem 1:
It seems that the RHEL 6 header file for getaddrinfo changed from
RFC 2553 to RFC 3943 and didn't provide any backwards compatibility. So
this meant that EAI_NODATA ceased to be defined. I searched around online
and found this link:
	http://osdir.com/ml/encryption.kerberos.devel/2005-02/msg00025.html
which gave a plausable workaround and I adopted it.

The above is implemented by the eai_nodata.patch file.

Problem 2:
In the src/appl/gssftp/ftpd codes, they implemented a function called
getline(), which apparently glibc (only) now defines with a different
signature.  I renamed this function to xgetline() instead.

The above is implemented by the getline.patch file.

This is for ticket #1998.