Fix recovery of cream jobs during stage-in. #2202
If the gridmanager restarts during stage-in of a cream job, the
gridmanager assumes stage-in has finished and tells cream to submit the
job to the back-end scheduler. This will lead to failures due to missing
or incomplete input files, which may not be detected at the CREAM/Condor
layer.

Now, the gridmanager sets StageInStart and StageInFinish at the
beginning and end of staging input, respectively. This lets it now
during recovery whether it needs to retry stage-in.

===VersionHistory=== Fix a bug that caused CREAM jobs to have incomplete
  input files if the condor_gridmanager crashed during transfer of those
  input files.