Add test cases for always running POST
===GT=== #2057
Test A here is supposed to succeed: the PRE Script fails,
so the node job is not run; but the POST script runs anyway.

Test B is supposed to fail: the test gives the -no_postrun
option on the command line, so when the PRE Script fails,
the whole node fails and the POST script does not run.