Fix Condor-C to blahp job submissions. #2248
The blahp expects the full path to the executable in the Cmd attribute.
When a job arrives via Condor-C or remote submit, Cmd has just the
basename of the executable, relative to Iwd. The gridmanager needs to
modify Cmd to the full path in the ad it gives to the blahp.

===VersionHistory=== PBS/LSF grid jobs that arrive in a queue via
  Condor-C or remote submission work properly again.