Get the maximum length supported by a deltacloud driver
In order to make the best decision about how to name an
instance, we need to know the maximum length supported
by a deltacloud driver.  This patch sets up gridmanager
to ask the deltacloud GAHP what that maximum is.  Once it
finds out the maximum, it passes that along to
build_instance_name, which will truncate the UUID as
necessary.
===GT=== #2264