Move create_failure into dcloudgahp_commands.cpp.
It never quite sat well with me that create_success was
in dcloudgahp_commands.cpp while create_failure was in
dcloudgahp_common.cpp.  It turns out that only one
place outside of dcloudgahp_commands.cpp was using it,
and using it very simply at that.  Just move it into
dcloudgahp_commands.cpp statically, and replace the
one other use with a hardcoded string.
===GT=== #2264