Replace the call to deltacloud_get_instance_by_id.
With the switch to libdeltacloud 0.8, an additional
deltacloud_get_instance_by_id() call was added to
the dcloud_start_worker() method.  Unfortunately, that
was the wrong thing to do; it is not guaranteed that
the instance ID can be looked up directly after the
deltacloud_create_instance() call.  To fix this,
generate a fake state of PENDING directly after we
return from the deltacloud_create_instance() call.
It will then be up to the upper layers (i.e. gridmanager)
to check in on the state.
===GT=== #2264