Move the JobRunning transition into ProcessInstanceAttrs.
Consider the following: a deltacloud EC2 job goes through the
following transitions:

GM_INIT -> GM_START -> GM_CLEAR_REQUEST -> GM_UNSUBMITTED ->
  GM_SAVE_INSTANCE_NAME -> GM_CREATE_VM -> GM_CHECK_AUTOSTART ->
  GM_SAVE_INSTANCE_ID -> GM_SUBMITTED -> GM_PROBE_JOB ->
  GM_SUBMITTED

At this point, the state machine exits while the EC2 instance
is booting.  The status will eventually be updated via BatchStatus.

However, when BatchStatus does see a state transition, it calls
StatusUpdate directly.  What happens then is that the job gets
set to RUNNING (via JobRunning()), and then the job is marked for
state re-evaluation.  The problem is that this is the wrong way
around; we need to re-evaluate the state *before* setting the job
to running, otherwise we never probe the now-running job for
attributes that external users might care about.

This patch moves the JobRunning() call into ProcessInstanceAttrs,
which means that you always have to get here through the state
machine.
===GT=== #2264