Fixed a problem causing dagman to fail to extract the dag node name from the submit event. #2335
This problem was introduced by my initial commit for #2335.
In the problematic patch, I replaced a call to fscanf() with a function
that read a single line, not realizing that the newline at the end of the
fscanf format string was causing it to read past the end of the line
until the next non-whitespace character on the following line.