Made DedicatedScheduler's SetAttributeString of ATTR_SCHEDULER to be NONDURABLE, #2367
The setting of ATTR_SCHEDULER need not be durable as it is reset
periodically if wrong. Making it DURABLE caused an fsync for every job
in every parallel cluster.

In fact, because the code to set the ATTR_SCHEDULER is (n^2-n)/2 in
the number of procs in a parallel cluster, the number of fsync calls
was dramatic.