Fix RETRY and jobstate
===GT===#2057
Had the wrong semantics, where all parts of node would have to succeed for the
node to succeed.  Also need to set retval before writing the jobstate log.

With this commit, all tests pass, except for job_dagman_always_run_post-{A,E},
which are testing for the wrong thing.