Make gt2_gahp safer with different Globus versions. #2389
We use a crude hack to optimize authentication of GASS server
connections which requires accessing private data structures of
globus_xio and globus_io, whose layout could change in future releases
of Globus. Now, we check that the version of these modules falls within
a known safe range. If they don't, we disable the optimization.